requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.15
openai>=1.12.0

# AWS dependencies
boto3>=1.34.0
botocore>=1.34.0

# Utility dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0

# Logging and monitoring
structlog>=23.2.0

# Development dependencies
pytest>=7.4.0
black>=23.12.0
isort>=5.13.0
mypy>=1.8.0
ruff>=0.1.9

fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart